requirements.txt
aiohttp==3.6.2
aiosqlite==0.13.0
async-timeout==3.0.1
asyncpg==0.20.1
attrs==19.3.0
beautifulsoup4==4.9.1
bs4==0.0.1
chardet==3.0.4
crontab==0.22.8
idna==2.10
iso8601==0.1.12
multidict==4.7.6
PyPika==0.37.11
soupsieve==2.0.1
tortoise-orm==0.16.13
typing-extensions==3.7.4.2
yarl==1.4.2
